Move LessonIntroModel
