+++ features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/lessons/intro/LessonIntroFragment.kt
@@ -28,6 +28,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.fragment.app.FragmentManager
+import app.ss.models.LessonIntroModel
 import com.cryart.sabbathschool.core.extensions.view.viewBinding
 import com.cryart.sabbathschool.lessons.R
 import com.cryart.sabbathschool.lessons.databinding.SsFragmentLessonIntroBinding
